Simplify mixins structure
